Remove long/double type support: initial commit

+++ modules/ballerina-core/src/main/java/org/ballerinalang/model/values/BFloat.java
@@ -24,9 +24,9 @@ package org.ballerinalang.model.values;
  */
 public final class BFloat extends BValueType {
 
-    private float value;
+    private double value;
 
-    public BFloat(float value) {
+    public BFloat(double value) {
         this.value = value;
     }
 
@@ -42,7 +42,7 @@ public final class BFloat extends BValueType {
 
     @Override
     public float floatValue() {
-        return this.value;
+        return (float) this.value;
     }
 
     @Override
@@ -57,7 +57,7 @@ public final class BFloat extends BValueType {
 
     @Override
     public String stringValue() {
-        return Float.toString(value);
+        return Double.toString(value);
     }
 
     @Override
